update
AddedContactDisplay on click gets you to the conversation that should be.

There is a "bug" where i put the domain two times. dont know where that is but on home, if we find the domain two times, it gets deleted so that there is only one time before sending the message.